Move audit log to backend folder
Fixes #7359
