+++ specifyweb/backend/stored_queries/execution.py
@@ -31,7 +31,7 @@ from .relative_date_utils import apply_absolute_date
 from .field_spec_maps import apply_specify_user_name
 from specifyweb.backend.notifications.models import Message
 from specifyweb.backend.permissions.permissions import check_table_permissions
-from specifyweb.specify.auditlog import auditlog
+from specifyweb.backend.workbench.upload.auditlog import auditlog
 from specifyweb.specify.models import Collectionobjectgroupjoin, Loan, Loanpreparation, Loanreturnpreparation, Taxontreedef
 from specifyweb.specify.utils import get_cat_num_inheritance_setting, log_sqlalchemy_query
