+++ specifyweb/backend/trees/views.py
@@ -20,7 +20,7 @@ from specifyweb.specify.field_change_info import FieldChangeInfo
 from specifyweb.backend.trees.ranks import tree_rank_count
 from . import extras
 from specifyweb.specify.api import get_object_or_404, obj_to_data, toJson
-from specifyweb.specify.auditcodes import TREE_MOVE
+from specifyweb.backend.workbench.upload.auditcodes import TREE_MOVE
 from specifyweb.backend.trees.stats import get_tree_stats
 from specifyweb.specify.views import login_maybe_required, openapi
