+++ specifyweb/backend/workbench/upload/upload.py
@@ -22,7 +22,7 @@ from jsonschema import validate  # type: ignore
 
 from specifyweb.backend.permissions.permissions import has_target_permission
 from specifyweb.specify import models
-from specifyweb.specify.auditlog import auditlog
+from specifyweb.backend.workbench.upload.auditlog import auditlog
 from specifyweb.specify.datamodel import Table
 from specifyweb.specify.func import Func
 from specifyweb.backend.trees.extras import renumber_tree, set_fullnames
